// src/app/core/models/Invoices.ts
export class DtInvoiceProduct {
    id: number;
    name: string;
    barcode: string;
    maxQuantity: number;
    quantity?: number;
    price: number;
    amount: number;
    discount: number;
    totalDiscount: number;
    netAmount: number;
}

export class DtInvoiceHighlights {
    totalItems: number;
    totalQuantity: number;
    totalAmount: number;
    totalDiscount: number;
    billDiscount: number;
    netAmount: number;
}

export class ReqInvoice {
    productsList: {
        id: number;
        quantity: number;
    }[];
    invoice: {
        billNumber: string;
        billDate: Date;
        remarks: string;
        referenceNumber: string;
        totalItems: number;
        totalAmount: number;
        totalDiscount: number;
        billDiscount: number;
        netAmount: number;
        paidAmount: number;
        amountReturned: number;
    }
    storeId: number;
}

// Generated by https://quicktype.io

export class RespInvoicesList {
    invoiceList: RespInvoice[];
    totalRecords: number;
}

export class RespInvoice {
    id: number;
    billNumber: string;
    remarks: null | string;
    billDate: string;
    totalItems: number;
    totalAmount: number;
    totalDiscount: number;
    billDiscount: number;
    netAmount: number;
    paidAmount: number;
    amountReturned: number;
    referenceNumber: string;
    listOfInvoiceItems: RespInvoiceItem[];
}

export interface RespInvoiceItem {
    id: number;
    quantity: number;
    barcode: string;
    price: number;
    cost: number;
    discountAmount: number;
    discountPercentage: number;
    netSalePrice: number;
    title: string;
    itemsId: number;
    tsClient: null;
}